Fix: Tests units
