+++ .github/workflows/deploy.yml
@@ -45,7 +45,7 @@ jobs:
         run: docker network create "microservice-network"
       
       - name: Set up MongoDB
-        run: docker-compose up -d
+        run: docker compose up -d
 
       - name: Check mongod
         run: docker ps
