+++ tests/integration/app.spec.ts
@@ -1,5 +1,5 @@
 import request from "supertest";
-import app from "../../src/server";
+import app from "../../src/index";
 
 describe("Transactions API", () => {
   it("should return a 200 status code", async () => {
