+++ tests/integration/getPurchasesIntegration.spec.ts
@@ -1,5 +1,5 @@
 import request from 'supertest';
-import app from '../../src/server';
+import app from '../../src/index';
 import { PurchasesRepository } from '../../src/repositories/purchases-repository';
 
 jest.mock('../../src/repositories/purchases-repository', () => {
